updates on navbar
