+++ src/common/components/Icon/set/spainFlag.jsx
@@ -1,0 +1,21 @@
+const spainFlag = ({ width, height, style }) => (
+  <svg
+    style={style}
+    width={width || '27px'}
+    height={height || '27px'}
+    viewBox="0 0 100 100"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <circle cx="50" cy="50" r="50" fill="#FFDA44" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M47.8542 0.040368C35.9211 0.534867 24.7331 5.17165 15.9801 13.2504C10.707 18.1173 6.57279 24.0078 3.76029 30.6615C3.47971 31.3253 3 32.5455 3 32.5954C3 32.6088 24.095 32.6198 49.8778 32.6198H96.7556L96.7265 32.526C96.6672 32.3354 96.1702 31.0813 95.938 30.5365C95.323 29.093 94.457 27.3034 93.7504 26.0156C91.1357 21.2506 87.7603 16.9334 83.7699 13.2504C77.1416 7.13269 69.0457 2.94266 60.2572 1.08145C56.261 0.235118 51.929 -0.128507 47.8542 0.040368ZM3 67.3941C3 67.4075 3.10567 67.6935 3.23483 68.0295C4.43213 71.1448 6.19796 74.5768 8.05217 77.3922C9.91617 80.2226 11.8828 82.6521 14.2693 85.0725C16.3786 87.2119 18.4322 88.9716 20.7917 90.6618C30.2483 97.4356 41.6906 100.669 53.375 99.8678C64.6867 99.0927 75.3982 94.466 83.7699 86.7391C89.011 81.9019 93.1226 76.0609 95.938 69.4531C96.1702 68.9082 96.6672 67.6541 96.7265 67.4635L96.7556 67.3698H49.8778C24.095 67.3698 3 67.3807 3 67.3941Z"
+      fill="#D80027"
+    />
+  </svg>
+
+);
+
+export default spainFlag;
