+++ src/common/components/Navbar/index.jsx
@@ -2,68 +2,68 @@ import {
   Box,
   Flex,
   IconButton,
-  Button,
+  Avatar,
   Stack,
   Collapse,
   useColorModeValue,
   useBreakpointValue,
   useDisclosure,
+  useColorMode,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverArrow,
+  Button,
 } from '@chakra-ui/react';
-import { useTranslation } from 'next-i18next';
-// import { useRouter } from 'next/router';
+import { useState, memo } from 'react';
+import { useRouter } from 'next/router';
+import PropTypes from 'prop-types';
 import NextChakraLink from '../NextChakraLink';
 import Icon from '../Icon';
-/* import Image from '../Image';
-import logo from '../../../../public/static/images/bc_logo.png'; */
 import DesktopNav from '../../../js_modules/navbar/DesktopNav';
 import MobileNav from '../../../js_modules/navbar/MobileNav';
+import Heading from '../Heading';
+import Text from '../Text';
 
-const NavbarWithSubNavigation = () => {
+import useAuth from '../../hooks/useAuth';
+import navbarTR from '../../translations/navbar';
+
+const NavbarWithSubNavigation = ({ haveSession }) => {
+  const router = useRouter();
+
+  const {
+    loginText, ITEMS,
+  } = navbarTR[router.locale];
+
   const { isOpen, onToggle } = useDisclosure();
-  const { t } = useTranslation(['navbar']);
+  const { colorMode, toggleColorMode } = useColorMode();
   const commonColors = useColorModeValue('white', 'gray.800');
+  const popoverContentBgColor = useColorModeValue('white', 'gray.800');
+  const commonBorderColor = useColorModeValue('gray.200', 'gray.900');
+  const { user, logout } = useAuth();
 
-  const EXTERNAL_ITEMS = [
-    {
-      label: t('menu.about-us'),
-      href: '/about-us',
-    },
-    {
-      label: t('menu.practice'),
-      href: '/interactive-exercises',
-    },
-    {
-      label: t('menu.read.title'),
-      icon: 'book',
-      description: t('menu.read.description'),
-      asPath: '/lessons', // For colorLink
-      subMenu: [
-        {
-          label: t('menu.read.child-1.label'),
-          // subLabel: t('menu.read.child-1.subLabel'),
-          href: '/lessons?child=1',
-        },
-        {
-          label: t('menu.read.child-2.label'),
-          // subLabel: t('menu.read.child-2.subLabel'),
-          href: '/lessons?child=2',
-        },
-        {
-          label: t('menu.read.child-3.label'),
-          // subLabel: t('menu.read.child-2.subLabel'),
-          href: '/lessons?child=3',
-        },
-      ],
-    },
-    {
-      label: t('menu.build'),
-      href: '/projects',
-    },
-    {
-      label: t('menu.bootcamp'),
-      href: 'https://4geeksacademy.com',
-    },
-  ];
+  const [settingsOpen, setSettingsOpen] = useState(false);
+
+  const closeSettings = () => {
+    setSettingsOpen(false);
+  };
+  const toggleSettings = () => {
+    setSettingsOpen(!settingsOpen);
+  };
+
+  const getImage = () => {
+    if (user && user.github) {
+      return user.github.avatar_url;
+    }
+    return '';
+  };
+
+  const getName = () => {
+    if (user && user?.first_name) {
+      return `${user?.first_name} ${user?.last_name}`;
+    }
+    return user?.github.name;
+  };
 
   return (
     <Box>
@@ -107,7 +107,6 @@ const NavbarWithSubNavigation = () => {
           />
           <NextChakraLink href="/" alignSelf="center" display={{ base: 'flex', md: 'none' }}>
             <Icon icon="logoModern" width="90px" height="20px" />
-            {/* <Image src={logo} width="30px" height="30px" alt="4Geeks logo" /> */}
           </NextChakraLink>
         </Flex>
 
@@ -117,12 +116,12 @@ const NavbarWithSubNavigation = () => {
           </NextChakraLink>
 
           <Flex display={{ base: 'none', md: 'flex' }} ml={10}>
-            <DesktopNav NAV_ITEMS={EXTERNAL_ITEMS} />
+            <DesktopNav NAV_ITEMS={ITEMS} haveSession={haveSession} />
           </Flex>
         </Flex>
 
         <Stack flex={{ base: 1, md: 0 }} justify="flex-end" direction="row" spacing={6}>
-          {/* <IconButton
+          <IconButton
             display={useBreakpointValue({ base: 'none', md: 'flex' })}
             _hover={{
               background: commonColors,
@@ -139,35 +138,142 @@ const NavbarWithSubNavigation = () => {
                 <Icon icon="dark" width="20px" height="20px" />
               )
             }
-          /> */}
-          <NextChakraLink
-            href="/login"
-            fontWeight="700"
-            fontSize="13px"
-            lineHeight="22px"
-            _hover={{
-              textDecoration: 'none',
-            }}
-            letterSpacing="0.05em"
-          >
-            <Button
-              display={useBreakpointValue({ base: 'flex', md: 'flex' })}
-              width="100px"
-              fontWeight={700}
-              lineHeight="0.05em"
-              variant="default"
+          />
+
+          {haveSession ? (
+            <Popover
+              id="Avatar-Hover"
+              isOpen={settingsOpen}
+              onClose={closeSettings}
+              placement="bottom-start"
+              trigger="click"
             >
-              {t('login')}
-            </Button>
-          </NextChakraLink>
+              <PopoverTrigger>
+                <Button
+                  bg="rgba(0,0,0,0)"
+                  alignSelf="center"
+                  width="20px"
+                  minWidth="20px"
+                  maxWidth="20px"
+                  height="30px"
+                  borderRadius="30px"
+                  onClick={() => toggleSettings()}
+                >
+                  <Avatar
+                    // name={user?.first_name}
+                    width="30px"
+                    marginY="auto"
+                    height="30px"
+                    src={getImage()}
+                  />
+                </Button>
+              </PopoverTrigger>
+
+              <PopoverContent
+                border={0}
+                boxShadow="dark-lg"
+                bg={popoverContentBgColor}
+                rounded="md"
+                width={{ base: '100%', md: 'auto' }}
+                minW={{ base: 'auto', md: 'md' }}
+              >
+                <PopoverArrow />
+
+                {/* Language Section */}
+                <Box
+                  width="100%"
+                  borderBottom={1}
+                  borderStyle="solid"
+                  borderColor={commonBorderColor}
+                  display="flex"
+                  justifyContent="space-between"
+                  padding="12px"
+                >
+                  <Text size="m">
+                    Language
+                  </Text>
+                  <Box display="flex" flexDirection="row">
+                    <Text display="flex" alignItems="center" gridGap="5px" size="sm" color="gray.600">
+                      <Icon icon="usaFlag" width="16px" height="16px" />
+                      ENG
+                    </Text>
+                    <Box width="1px" height="100%" background="gray.350" margin="0 6px" />
+                    <Text display="flex" alignItems="center" gridGap="5px" size="sm" color="gray.600">
+                      <Icon icon="spainFlag" width="16px" height="16px" />
+                      ESP
+                    </Text>
+                  </Box>
+                </Box>
+
+                {/* Container Section */}
+                <Box p={4}>
+                  <Stack gridGap="10px" pb="15px">
+                    <Flex alignItems="center" gridGap="6px">
+                      <Box as="span" fontSize="18px" lineHeight="18px">
+                        Welcome
+                      </Box>
+                      <Heading as="p" size="18px">
+                        {getName()}
+                      </Heading>
+                    </Flex>
+
+                    <Flex alignItems="center" gridGap="6px">
+                      <Box as="span" fontSize="18px" lineHeight="18px">
+                        Current Role:
+                      </Box>
+                      <Heading as="p" size="18px">{`${user?.roles[0].role}`}</Heading>
+                    </Flex>
+                  </Stack>
+
+                  <Flex padding="20px 0" alignItems="center">
+                    <Button gridGap="10px" onClick={logout} width="100%" py="25px">
+                      <Box as="span" fontSize="15px">
+                        Logout
+                      </Box>
+                      <Icon icon="logout" width="20px" height="20px" />
+                    </Button>
+                  </Flex>
+                </Box>
+
+              </PopoverContent>
+            </Popover>
+          ) : (
+            <NextChakraLink
+              href="/login"
+              fontWeight="700"
+              fontSize="13px"
+              lineHeight="22px"
+              _hover={{
+                textDecoration: 'none',
+              }}
+              letterSpacing="0.05em"
+            >
+              <Button
+                display={useBreakpointValue({ base: 'flex', md: 'flex' })}
+                width="100px"
+                fontWeight={700}
+                lineHeight="0.05em"
+                variant="default"
+              >
+                {loginText}
+              </Button>
+            </NextChakraLink>
+          )}
         </Stack>
       </Flex>
 
       <Collapse in={isOpen} animateOpacity>
-        <MobileNav NAV_ITEMS={EXTERNAL_ITEMS} />
+        <MobileNav NAV_ITEMS={ITEMS} haveSession={haveSession} />
       </Collapse>
     </Box>
   );
 };
 
-export default NavbarWithSubNavigation;
+NavbarWithSubNavigation.propTypes = {
+  haveSession: PropTypes.bool,
+};
+NavbarWithSubNavigation.defaultProps = {
+  haveSession: false,
+};
+
+export default memo(NavbarWithSubNavigation);
