+++ styles/theme.js
@@ -56,6 +56,7 @@ const CustomTheme = extendTheme({
       100: '#EDF2F7',
       200: '#E2E8F0',
       300: '#CBD5E0',
+      350: '#C4C4C4',
       400: '#A0AEC0',
       500: '#718096',
       600: '#606060',
